KafkaLab/backend/consumer/requirements.txt
kafka-python
scikit-learn
joblib
nltk
